Move user-facing exceptions to UI layer
User facing exceptions include localisable strings, that are only available
at the UI layer.
