+++ GitUI/NBugReports/UserExternalOperationException.cs
@@ -1,6 +1,8 @@
 ﻿#nullable enable
 
-namespace GitCommands
+using GitCommands;
+
+namespace GitUI.NBugReports
 {
     /// <summary>
     /// Represents errors that occur during execution of user-configured operation, e.g. a script.
