+++ GitUI/Script/ScriptRunner.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Windows.Forms;
 using GitCommands;
 using GitUI.HelperDialogs;
+using GitUI.NBugReports;
 using GitUIPluginInterfaces;
 
 namespace GitUI.Script
